include entrypoints in Python get_exporter APIs
not just in the application
